Implement GetSchema. (#1181)
These changes implement the resource provider's GetSchema method in
order to facilitate code generation (e.g. for docs and for the proposed
import command).

As part of these changes, I refactored the Makefile s.t. each language
has its own target, and I updated the generator s.t. the schema is read
from disk unless it is the target language. The only exception is the
.NET code generator, which still uses the Swagger spec for YAML support.

Fixes #1180.