Update json-patch mod to fix hangs on pulumi update (#1223)
* Update json-patch mod to fix hangs on pulumi update

A performance bug was discovered that can cause CRD
diffs to fail. This bug is fixed in an open PR [1] to the upstream
repo, but has not been merged yet. In the meantime, I forked
the json-patch repo, and applied the fix [2]. This change will be
reverted once the upstream patch merges, and the module will
be updated to point at the fixed changeset.

[1] https://github.com/evanphx/json-patch/pull/108
[2] https://github.com/pulumi/json-patch

* gofmt