Set plugin version for Go SDK invoke calls (#1325)
For invoke calls, the version must be set to avoid using
an older version of the plugin that does not support the
specified invoke function. We don't currently have a way
of automatically determining the version [1] to set for the
Go SDK, so we're hardcoding the version to the v2.6.1
plugin for now. This version contains the latest
implementation for all of our provider invoke functions.

Note that all of our other SDKs are already setting the
auto-discovered version in similar invoke calls.

[1] https://github.com/pulumi/pulumi-kubernetes/issues/1324