Moving to a command dispatch workflow for PRs
Fixes: #1341

This will mean that when a PR is opened, linting and an initial
build of the binaries will be performed to ensure there are no
obvious error.

The PR will have a comment added to it suggesting that the user needs
a maintainer to run the acceptance tests. The maintainer will run
the command `/run-acceptance-tests` and will wait for the results

This will trigger a full end-to-end set of tests including spinning
up the test infrastructure.

It is super important that the PR is checked that no secrets etc are
leaked before running the test command

Please note, as this is a command dispatch, this cannot be tested from
a PR - it must be in master and we can interate there!