Add support for Create/Update previews. (#1355)
Implement support for the recently-added "provider preview" feature of
the Pulumi plugin interface that allows a resource provider to return
known values from a call to Create or Update during a preview. This
feature improves the usability of previews by allowing more values to be
treated as known.

When previewing a Create or Update operation, the Kuberentes provider
will either return the result of a dry-run performed by the API server
or will return the inputs as-is. The dry-run is preferred, but is only
possible when all of the inputs to the resource are known, the API
server supports dry-running the resource in question, and the user has
explicitly enabled the feature using the `enableDryRun` configuration
property. Returning the inputs as-is matches the behavior previously
implemented by the Pulumi engine.