Error on delete if cluster is unreachable (#1379)
We previously considered a resource to be deleted if the
cluster is unreachable during an update. The intent was to
gracefully handle cases where the underlying cluster was
deleted (perhaps as part of another stack). Unfortunately,
this change led to cases where resources were erroneously
deleted from the state due to network partitions.

This change reverts that behavior so that an error is returned
for resource deletion if the cluster is unreachable.