Skip Helm test hook resources by default (#1467)
Many Helm charts include test hooks that can be
used by the "helm test" command to validate an
installation, but should not be included as part of
a default chart installation. This change disables
the creation of test resources by default, and includes
a new Chart parameter that can be used to override
this default and install the test resources.