Change k8s API removals from error to warning (#1475)
The provider checks for deprecated and removed API
versions of k8s resources to provide clearer messages
for affected APIs. These messages are intended to be
informational, and should not block the deployment of
a resource if the cluster accepts it. To avoid inadvertently
failing a deployment when the upstream removal version
has changed, print a warning rather than returning an error.