Error on refresh for an unreachable cluster.
Previously, the refresh action would assume that an unreachable
cluster meant that the resources had been deleted out of band,
and would delete them from the state with a warning. This deletion
only affected the Pulumi state, not the actual k8s resource, and
could be repaired by reverting to the previous checkpoint.

Unfortunately, this behavior can lead to unintended state changes
in the case of a network partition, especially if the refresh operation
is running as part of an automated job.

This commit changes the default behavior of refresh to return an
error in the case of an unreachable cluster. This change means that
the state may have to be updated manually if the cluster was actually
deleted, but this seems like a preferable alternative to unexpected
state changes.