[helm] Don't pass opts to child resources (#1539)
The `opts` passed to a component should be passed to it's `super` call, but not to it's children.  Child resources automatically inherit those options that they should from their parent - so the resource options applied to a parent should not also be passed down to cildren explicitly.

In 3 of the 8 implementations of Helm Chart in this codebase, we were passing the component opts to the children.  In the other 5 we were (correctly) not passing them.

Fixes #1538