[sdk/go] Re-add deprecated Provider file
The Go SDK erroneously included an old version of
the Provider resource until v3.6.0 when it was incidentally
removed as part of another update. To avoid breaking
user code that was referring to this path, re-add a copy
of the Provider with included deprecation notices to
move to the intended path.