Fix DaemonSet name on diff
The DaemonSet name was set with a lower case s (Daemonset), but the kind
in the state is created with upper case (DaemonSet), and this was
preventing changes on the .spec.selector to trigger a replace.

This issue was also causing the Kubernetes API to fail as this field is
immutable, but pulumi was trying to update it.