Use go embed rather than generating a .go file (#2021)
The newer go versions (>=1.16) include the ability to directly embed
files as variables, so it's no longer necessary to generate a .go file
with the contents of another file. It _is_ still necessary to have a
generation step to rewrite the schema itself. This is how the boilerplate provider build is structured now:

    https://github.com/pulumi/pulumi-provider-boilerplate/blob/main/provider/cmd/pulumi-resource-xyz/main.go#L26

Since I'm porting that change, I've included a little failsafe in the
Makefile for when people start using the changed build. The build would
fail anyway, but it's less mysterious this way.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>