Change .metadata.name to optional for all Patch resources
This is a minor breaking change that updates the `.metadata.name` field from required to optional for all Patch resources. This resolves an issue where a nested reference to the ObjectMeta type also required the name field to be set (e.g., for Deployment, StatefulSet, etc.). Instead of requiring the name field in the SDKs, this requirement will be enforced by the provider. This change trades some clarity in the API usage for semantic correctness.