Fix field names in CRD schemas (#2128)
Several fields in the CRD schema include dashes in the key name, which cause problems with SDK code generation. As a workaround, these keys were renamed in the SDKs to use underscores rather than dashes. However, the provider was incorrectly using the underscore conversion in the API call to the k8s cluster. This incorrect behavior was uncovered using the Server-Side Apply mode, but was present in both modes. The provider now correctly converts these keys back to the correct form before calling the k8s API.