Skip computing a preview for Patch resources
Since Patch resources do not have to specify a complete resource, attempting to patch a nonexistent resource will return an Invalid error response during preview. To avoid this, skip computing a preview for Patch resources, and return the inputs directly. This is the same behavior used for resources containing unknowns.