Enable Server-Side Apply mode by default (#2206)
* Fix failing test

This test was accessing an Output property without using an apply, and the value had not resolved already. Apparently this was working with the legacy Client-Side Apply implementation, but the test code was unsafe.

* Fix replacement behavior for immutable fields

With Server-side Apply previews, updating an immutable field in a resource will result in a 422 error code from the API server. In these cases, Pulumi will replace the resource, so we need to ignore the error rather than failing the preview.