Qualify Resource type from Pulumi namespace
The next release of Kubernetes (1.26.0) has an API named `resource`. In
the .NET codegen, this will end up being a namespace `Resource`,
clashing with the type name `Resource` which is implicitly in scope.

To avoid this, give the Pulumi namespace an alias and use that to
qualify the `Resource` type.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>