Only run setup for the language under test
The build_sdk and test jobs both use a matrix to run a similar series of
steps for each language (Java, go, etc.). But it's not necessary for
every step to run for every language; in particular, it's not necessary
to set up any language other than the one under consideration, and Go at
the set version to run the tests.