Use a kind config to specify Kubernetes version
Some of the tests (I'm looking at you, TestDotnet_CustomResource) use
APIs which have been removed from recent versions of Kubernetes
(respectively, apiextensions/v1beta1, removed in Khbernetes v1.22 August
2021). The SDK supports deprecated and removed APIs, so it's valid to
use a removed version; but, it means we need to run an old cluster for
the test to pass.

Obviously there's a tension between testing old APIs and testing with
up-to-date clusters -- I've erred on the side of not changing everything
at once.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>