Move language versions to env entries
The versions of each language compiler/runtime are given in matrix
entries in individual steps. This produces the right results, but it
makes the workflow structure confusing and obscures what's actually
running (since each matrix value is put in the job name).

Since there's only one version used for each language, and it's the same
for every step, they may as well be workflow-level env entries.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>