Always set a field manager name (#2271)
Recent versions of Kubernetes have started to create a managedFields entry for every resource, even those that use Client-Side Apply (CSA). The name of the field manager is inferred from the name of the executable performing the update. This can lead to cross-platform conflicts, because the executable name can be different (e.g., myprogram vs. myprogram.exe). This leads to multiple field managers owning the same fields.

To avoid auto-naming conflicts like this, always set a field manager name. In the CSA mode, this name will always be "pulumi-kubernetes". In Server-Side Apply (SSA) mode, the name will be "pulumi-kubernetes-<suffix>", with a randomly-generated suffix.