Expect provider version to be supplied
Taking a leaf from pulumi-azure-native: instead of trying to calculate
the provider version in the Makefile, expect it to be passed in the
environment or as a Makefile variable, with a (valid) dummy value by
default.

This means the CI can calculate the version for use elsewhere (e.g., for
publishing), and be sure it's the same value used here.

Also: use pulumictl to convert the provider version to each of SDK
versions. To make sure a pulumictl with convert-version is used,
download it as part of `make ensure`.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>