Fix await status for Job and Pod (#2299)
Fix a regression where the intermediate status of Job and Pod resources was no longer printed during an update.

The cloud-ready-checks package changed the way that messages are populated in the state checker. Previously, each condition was returned as a StatusMessage when the Update method was called. During the refactor, the Update method was replaced by the ReadyDetails method, which changed the return type. The updated return type includes a Description and an optional message that was only set when a problem was detected. The regression was caused by failing to print the description if the message was empty. Thus, intermediate status was still printed in case of errors, but was erroneously skipped on the happy path.

This change updates the two usages of this library to iterate the result objects and log each result as a status message. This approximates the previous behavior, and fixes the regression. The cloud-ready-checks package was intended to make this process easier, but the current interface still needs some work, since it shouldn't be necessary for clients to understand this nuance between the Description and Message field.