Factor out build and test pipelines
This commit factors out the build and test bits of workflow, which are
used at least by run-acceptance-tests and (main branch) build. Since
these were generated, they have the same content (but I eyeballed them
side by side anyway, to check). I reapplied some changes I made in prior
commits, to work with the Makefile (and fixed up a missing phony target
used in the workflow).

The factored-out workflow takes a git ref, being the commit at which to
clone, since each entry point (a workflow that's triggered, rather than
called) will have to calculate and supply it.

This commit tries not to change anything else, so e.g., acceptance tests
use the head of the PR branch rather than the merge result which is odd
and likely to be changed in subsequent commits.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>