Update all CRDs to v1 versions
Kubernetes v1.22 dropped support for v1beta1 CRDs. The test cluster for CI uses GKE, which just dropped support for v1.21 clusters. This change updates all tests to use v1 CRDs, and upgrades the version of the CI cluster.

Kubernetes v1.21 reached EOL in June 2022, so drop the CRD upgrade test as well, since this is not supported on current cluster versions.