Add HTTP request timeout option to KubeClientSettings
The HTTP client used to communicate with the Kubernetes server has a default timeout of 32 seconds. In some cases, this timeout is too low, and causes failures during provider initialization. Add an option to override this timeout value. The timeout value can be overridden in the following ways, in order of precedence:

1. The `kubeClientSettings:timeout` parameter for the Provider
2. The PULUMI_K8S_CLIENT_TIMEOUT environment variable

Note that this is distinct from the timeouts resource option, which controls how long Pulumi will wait for a resource operation to complete.