Handle resource change from static name to autoname under SSA (#2392)
Server-side apply mode had a bug where changing a resource from having a static `.metadata.name` to being autonamed by removing the entire `.metadata` property would return an error. This change updates the diff logic to account for this case, which will result in the resource being replaced as expected.