Drop support for Kubernetes clusters older than v1.13 (#2414)
As part of upgrading the client-go dependency, we need to drop support for Kubernetes versions older than v1.13.0 (released December 2018, EOL as of October 2019). client-go has dropped support [1] for the dry-run verifier because all cluster versions v1.13 and greater include this support. As a result, we drop the dry run check from the provider, and enforce a minimum version of v1.13.

Additionally, one of the options for kustomize support moved [2], so update the calling code to match.

[1] https://github.com/kubernetes/kubernetes/pull/114294
[2] https://github.com/kubernetes-sigs/kustomize/issues/4945