[internal] Add function to prune map keys
Add a "pruneMap" function and supporting "pruneSlice" function to prune map keys that are not present in a target map.

Kubernetes sets a significant amount of additional state on cluster resources after they are created, which complicates diffing with the subset of properties that are set by a user program. The pruneMap function can be used to preprocess live resource state before diffing against program inputs.