Drop support for legacy pulumi.com/initialApiVersion annotation (#2443)
The pulumi.com/initialApiVersion annotation was briefly used to store information about the apiVersion used to create each resource in early versions of the provider (v1.x). The logic was updated to store this information in the Pulumi state rather than directly on the resource. Subsequent versions of the provider continued to support this annotation as a fallback option in case the state did not contain this information.

Drop support for the annotation to simplify the provider logic.