Fix StatefulSet await logic for OnDelete update  (#2473)
The await logic for StatefulSets previously only worked for the default RollingUpdate strategy. StatefulSets that were updated using the OnDelete strategy would incorrectly report that they were not ready because the provider was not accounting for differences in the status fields that depend on the update method. This change fixes the await logic so that both update strategies correctly report readiness.