Skip wait for Pods on headless Service (#2475)
The await logic for Service resources normally waits for selected Pods to be ready before marking the Service ready. This doesn't work correctly for StatefulSet resources, which use a "headless" Service (i.e., Service without a ClusterIP) to reference its Pods. This led to a circular dependency if the StatefulSet referenced the Service name in the Pulumi program, because the StatefulSet would not be created until after the Service was ready, but the Service was waiting on the StatefulSet to create Pods. This change fixes this dependency by skipping the Pod check for any headless Service, rather than only headless Services that don't have Pods selected.