Improve deleteUnreachable workflow (#2489)
The provider includes a deleteUnreachable option that will delete unreachable cluster resources from state. This was added as an opt-in behavior to avoid accidental state deletion for a temporarily unreachable cluster. This change improves this workflow in two ways:

1. The error messages for both the refresh and destroy cases were revised for clarity, and suggest using the PULUMI_K8S_DELETE_UNREACHABLE environment variable to enable this behavior.
2. The destroy workflow was changed to succeed for unreachable resources when the deleteUnreachable option is set rather than requiring the user to refresh the state as a separate step.

Note that this behavior can also be enabled with the Provider's deleteUnreachable configuration, but we do not mention this in the error message since it could lead to undesired state changes in case of a temporarily unreachable cluster.