Improve error handling for List resources (#2493)
List resources are included in the SDK for compatibility with various YAML manifests and Helm charts. These resources are not currently able to be created directly, but the failure condition was not handled explicitly. Instead, users would get an error when the provider failed to determine if the resource was namespaced. This is fixed with the following changes:

1. Create a lookup table for known List resources
2. Return an error during Check for List resources with a descriptive error message

This change also adds a few missing kinds to the Namespaced lookup table that have been added since the table was created initially.