Fix diff for CRD .spec.preserveUnknownFields
The v4 diffing logic relies on resource normalization to accurately compare input values to live state. Most resources are normalize by converting the Unstructured resource to its typed equivalent and then back, but this process requires a defined conversion scheme. However, CRDs do not have a conversion scheme defined.

CRD resources have a deprecated parameter ".spec.preserveUnknownFields" that the apiserver removes from the created CRD if the value is false. This causes problems for diffing, because the inputs and the live state are continuously out of sync.

This change adds CRD-specific normalization logic to avoid causing a continuous diff if this field is set in the resource definition.