Fix diff for resources modified by the cluster
In some cases, Kubernetes modifies resources before they are created on the cluster. One example is for Secret resources: if the `.stringData` field is set, it will be rewritten as a base64-encoded value into the `.data` field, and `.stringData` will be removed.

The updated diff logic introduced in TODO was incorrectly reporting a change on every update for these cases. This fix updates the input pruning logic used by Diff to account for cases where the input fields have not changed, but are not present in the live state.