Fix diff for CRD with status set on input (#2512)
* Fix continuous diff for Secret stringData field

If the `stringData` field is set on a Secret resource, then Kubernetes will automatically replace it with an equivalent `data` field. This is causing a continuous diff with the updated logic from 6e329f3. This change updates the input diff logic to include this rewrite.

* Improve normalization and test

* Fix diff for CRD with status set on input

The CRD status field is an output field, but the apiserver will ignore it if set. This is common with CRDs generated from helm charts. Preemptively remove the status field to avoid a continuous diff in this case.

* Update changelog for v4.0.2 release