[internal] Fix parallel test flake
Two of the examples tests install different versions of the same CRD. This works in CSA mode because they overwrite each other, but cause a conflict in SSA mode. Since CRDs are cluster-scoped, it's nontrivial to fix since the tests run in parallel. Explicitly use CSA mode until we refactor the tests to run on separate clusters.