Improve resource normalization for diffing
The logic for normalizing Unstructured resources was inadvertently modifying resources in place because the argument was a pointer type. Change the logic to make a copy of the resource and then return a normalized version of the copy. This avoids returning a partially-modified version of the resource if normalization fails partway through the process.

As part of this change, the following updates were made:
1. Make a copy of the Unstructured resource rather than modifying the original directly.
2. Remove error return and related handling since normalization returns the input resource in case of failure.
3. Proactively remove output-only metadata fields as part of normalization.
4. Remove post-normalization pruning step. This was erroneously pruning identical input and output due to the pointer type, and is no longer needed.
5. Fix invalid CRD test data.