Fix helm.v3.Release replace behavior (#2532)
The diff logic for the helm.v3.Release resource was intended to replace the resource if the name or namespace changed. However, the names of the fields were incorrect, so the provider was erroneously choosing to update rather than replace the resource. Update the fields to the correct names so that the resource is replaced as intended.