Add codecov to this repo (#2550)
This adds basic code coverage for our provider binaries as exercised by our unit tests. An example of the comment produced by the tool occurs on this PR; the coverage report is informational only (will not block PR checks).

Details
Generates a coverage report while running the test target; adds a ci step to upload the coverage report during PR acceptance tests and master builds

We don't want to over index on this metric - our more interesting issues arise from the interactions with the upstream service - however, this is a simple add and gives us a little bit of data we can use to encourage better quality.