Helm Release detect changes to local chart (#2568)
### Proposed changes

This PR fixes a usability problem with the
`kubernetes:helm.sh/v3:Release` resource when using a local chart. With
this fix, Pulumi automatically detects changes to the rendered output of
the chart and performs a Helm upgrade accordingly. It does this by
persisting a SHA256 checksum of the chart output into the Pulumi stack
state.

Users may suppress change detection with the `ignoreChanges` option,
i.e. `ignoreChanges: ["checksum"]`.

**Upgrade considerations**: after upgrading the provider, users should
expect to see a diff (`checksum` property) and a corresponding Helm
upgrade. Helm upgrades are generally idempotent.

### Related issues (optional)
Closes #2118