Gracefully handle ignoreChanges without user needing to refresh a stack (#2566)
Note: this is stacked on top of
https://github.com/pulumi/pulumi-kubernetes/pull/2570

This PR updates the logic for handling fields added to ignoreChanges.
Previously, to accurately ignore a field without hitting an SSA conflict
error, the Pulumi user would need to run a refresh BEFORE running
`pulumi up` or `pulumi preview`. This is due to the provider using the
last known value of the ignored field stored in state when computing
what should be set in the ignored field when the resource is sent to the
k8s API. If drift has occurred on that field and is not reflected in the
Pulumi state as it is out of date, then the provider sends old
information to the API server resulting in a field conflict error.

As we hit the live server to obtain the resources' up to date
`resourceVersion`, we can also use this API result to gracefully handle
fields marked within ignoreChanges. This PR uses the fieldpath package
from upstream k8s libraries to parse the serialized json managed fields
entries and compares it to a normalized version of the Pulumi program's
ignoreFields value.

Fixes: #2542