Fix normalizing fields with empty objects/slices (#2576)
### Proposed changes
This PR further improves the resource normalization logic.

Prior to this merge, field values that consists of an empty
`map[string]interface{}` or empty slice would be discarded. This means
that normalizing `{parentField: {childField: {} }}` would result in the
childField being unset as the empty object (`{}`) value for `childField`
is discarded. In Kubernetes resources, especially when defining
NetworkPolicies and ingress/egress rules, `{}` != unset field.

A new test case is added to validate that normalizing NetworkPolicy
resources does not result in unwanted behaviour. To do this, the GKE
cluster we spin up for testing also enables the Calico network
enforcement.

### Related issues (optional)

Fixes: #2538