Propagate cancellation to Helm Release provider action (#2579)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes
Closes #2485.

This PR implements cancelation for `Create` and `Update` operations on
the Helm Release resource. When the user presses ctrl-c, the Helm
operation is cancelled and the stack state is updated accordingly. This
is an improvement over the current behavior where the Helm release may
be orphaned by Pulumi.

A future enhancement would be to implement cancelation for the `Delete`
operation, once it becomes possible in the Helm v3 library (see
[PR](https://github.com/helm/helm/pull/12109) and the `RunWithContext`
variants).

_Note: no new tests are provided because I don't see a way to simulate
cancelation using the integration test framework._

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->
Related to https://github.com/pulumi/pulumi/issues/14057. It is safe to
merge this PR beforehand.