Fix for Helm Import Regression (#2605)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->


### Proposed changes

This is a series of commits to fix a p1 and to improve the import of
Helm Releases. The regression occurred because `Read` attempted to use
the provider-generated inputs to compute a checksum, and returned an
error if it couldn't, and this caused the `pulumi import` tool to fail
in cases where the inputs aren't guessable. The fix is to eat the
problem during `Read`.

While implementing the fix, it became apparent that the templating code
for the `Release` resource should be decoupled from the `Chart`
resource, mainly to improve the consistency of how chart names are
handled.

Specific changes:
- suppress error in computing inputs during `Read`.
- decouple the templating code.
- simplify the chart linting code.
- use a timeout that is consistent with the Delete op.
- new tests for the import variants.

<!--Give us a brief description of what you've done and what it solves.
-->
Closes https://github.com/pulumi/pulumi-kubernetes/issues/2596

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

---------

Co-authored-by: Ramon Quitales <ramon@pulumi.com>