Improved search functionality for Helm Import (#2610)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes
This PR enhances the `Release` resource to search harder for a matching
chart when importing a release. When Helm installs a chart, it doesn't
record any information about the chart location, simply its name. Pulumi
attempts to locate a chart:
1. (existing) by looking for an expanded chart directory (e.g. `nginx/`)
in the program directory.
2. (new) by looking for a chart archive in the program directory, with
or without the version info (e.g. `nginx-15.3.4.tgz`).
3. (new) by looking in the local Helm repositories for a chart, matching
the name and version.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->